backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "chromadb==0.5.0",
    "fastapi==0.110.0",
    "groq==0.9.0",
    "httpx==0.27.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "langsmith>=0.2.0",
    "pydantic>=2.7.4",
    "pydantic-settings>=2.4.0",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]>=3.5.0",
    "sentence-transformers==2.7.0",
    "sqlalchemy==2.0.29",
    "uvicorn[standard]==0.29.0",
]
